requirements.txt
pyyaml==6.0.1
markdown==3.5
jinja2==3.1.2 